Added Lithuanian keyboard pack.
This is based on the original Lithuanian pack still available on Google Play, but vastly expanded.
In addition to the original three-row QWERTY layout, this pack now also ships:
- four-row QWERTY-based layout (similar to the default layout on all PCs)
- three-row ĄŽERTY-based layout (similar to the official Lithuanian standard keyboard)
- 16-key vertical layout (which switches to the the aforementioned four-row layout in landscape mode)

I have also cleaned up some files which are generated at compile time and added them to .gitignore.

On top of all that, I am adding a crowd-sourced autocorrection dictionary. That one needs some reviewing, but I think it should be a good start.

The screenshots under StoreStuff were taken on my Samsung Galaxy S2 phone. I couldn't find the theme that was used for the original screenshots, so the ones I made are quite dark and might need to be re-taken.