Compact Dvorak layout
A keyboard for Dvorak users with small screens or fat fingers.
This layout preserves some strengths of the Dvorak layout. It provides
easy access to the most common keys, and common bigrams are dispersed.
At the same time, this keyboard is recognizable (once you see what is
happening) for those familiar with the Dvorak layout.