Speedup and improve character lookups
We now use a lookup mapping each character to a key.
This greatly improves the speed of path generation.
Additionally, we now try looking for matching keys for
special characters, and only afterwards look for the base
character. E.g., previously, the "ä" key was not used for path
generation, but instead "a" (its base character) was used.