Fix a glitch in auto-dictionary
Fixing a mistake of mine. While set to "Do not learn new words", ASK didn't promote automatically learned words to user dictionary, but still considered them promotion candidates and counted the frequency of use.

This means that if a new word was typed more than 9 times, ASK is set to "Learn after some use", and it's typed once more, it's still promoted because it was typed 10 times total.

With this new fix, "Do not learn new words" truly means no new words will be learnt.