coptic: Add initial dictionary.
This wordlist is based on the dictionary published by the
Coptic Scriptorium project, along with information extracted
from some other documents.

It's rather ineffective at this point, in part because the
corpus is limited and needs custom ingest scripts, and in
part because the synthetic nature of Coptic makes a compact
dictionary of complete words impossible.

Nevertheless it constitutes a starting point.