Fix cursor bug
This bug was visible when swiping up to use the keyboard extension. This keyboard extension was automatically moved so it would be completely visible on the sscreen, but the cursor would start at the number 1, and remain at a fixed distane of the finger.

This meant that if you swiped up on the sixth key of a given row, the rightmost symbol you could reach was the 5, and releasing the finger anywhere in the left hand would select 1 or ./gradlew googleJavaFormat